Fix deprocation warning
node_redis: Deprecated: The EVAL command contains a "null" argument.
This is converted to a "null" string now and will return an error from v.3.0 on.
Please handle this in your code to make sure everything works as you intended it to.

see https://codio.myjetbrains.com/youtrack/issue/codio-8251